Add files via upload
A recreation director needs to keep track of summer T-ball teams, games, and season records. He has
hired you to write a computer program to make his life easier. You have created a menu driven program
that allows the director to use a database to easily manipulate teams, games, and scores. It also allows the
database to be refreshed for a new year.